Dynamically get instance network limits

+++ aws/allocate.go
@@ -39,7 +39,7 @@ func (c *allocateClient) AllocateIPsOn(intf Interface, batchSize int64) ([]*Allo
 		NetworkInterfaceId: &intf.ID,
 	}
 
-	available := int64(c.aws.ENILimits().IPv4 - len(intf.IPv4s))
+	available := c.aws.ENILimits().IPv4 - int64(len(intf.IPv4s))
 
 	// If there are fewer IPs left than the batch size, request all the remaining IPs
 	// batch size 0 conventionally means "request the limit"
@@ -109,7 +109,7 @@ func (c *allocateClient) AllocateIPsFirstAvailableAtIndex(index int, batchSize i
 		if intf.Number < index {
 			continue
 		}
-		if len(intf.IPv4s) < limits.IPv4 {
+		if int64(len(intf.IPv4s)) < limits.IPv4 {
 			candidates = append(candidates, intf)
 		}
 	}
